--- src/screens/RegistrationScreen.tsx
@@ -1,79 +1,0 @@
-import React, { useState, ReactElement } from 'react'
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native'
-import {
-  useFonts,
-  Merriweather_400Regular,
-  Merriweather_700Bold,
-} from '@expo-google-fonts/merriweather'
-import { AppLoading } from 'expo'
-import LabeledInput from '../molecules/LabledInput'
-
-function RegistrationScreen(): ReactElement {
-  const [userName, setUserName] = useState('')
-  const [userEmail, setUserEmail] = useState('')
-  const [userPassword, setUserPassword] = useState('')
-
-  const [fontsLoaded] = useFonts({
-    Merriweather_400Regular,
-    Merriweather_700Bold,
-  })
-
-  if (!fontsLoaded) {
-    return <AppLoading />
-  }
-
-  const onSubmit = () => {
-    console.log(userName, userEmail, userPassword)
-
-    // validation ? pass : highlight mistakes, return
-    // fetch ? response.ok : response.error, return
-  }
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.header}>Регистрация</Text>
-      <LabeledInput label='ФИО' onChangeText={setUserName} />
-      <LabeledInput label='Адрес эл. почты' onChangeText={setUserEmail} />
-      <LabeledInput
-        label='Пароль'
-        onChangeText={setUserPassword}
-        secureTextEntry
-      />
-
-      <TouchableOpacity onPress={onSubmit} style={styles.buttonStyle}>
-        <Text style={styles.buttonText}>Зарегистрироваться</Text>
-      </TouchableOpacity>
-    </View>
-  )
-}
-
-const styles = StyleSheet.create({
-  header: {
-    fontSize: 30,
-    fontFamily: 'Merriweather_700Bold',
-    color: '#C23232',
-  },
-  container: {
-    marginTop: 136,
-    flex: 1,
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-  buttonStyle: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginTop: 60,
-    width: 320,
-    maxHeight: 56,
-    backgroundColor: '#C23232',
-    borderRadius: 10,
-  },
-  buttonText: {
-    fontSize: 20,
-    color: '#fff',
-    fontFamily: 'Merriweather_400Regular',
-  },
-})
-
-export default RegistrationScreen
